make sure feedform norm output in TF block is casted to the same dtype as input dtype

+++ mistral_jax/one_file_ref.py
@@ -195,7 +195,7 @@ class TransformerBlock(eqx.Module):
         normed_x = jax.vmap(self.attention_norm)(x.astype(jnp.float32)).astype(x.dtype)
         r = self.attention(normed_x, cos_freq, sin_freq, positions, mask)
         h1 = x + r
-        h2 = jax.vmap(self.ffn_norm)(h1.astype(jnp.float32)).astype(jnp.float16)
+        h2 = jax.vmap(self.ffn_norm)(h1.astype(jnp.float32)).astype(h1.dtype)
         h2 = jax.vmap(self.feed_forward)(h2)
         out = h1 + h2
         return out
